Move CPT migrations to v1.4.0
