+++ classes/class-plugin-migrations.php
@@ -11,7 +11,7 @@ namespace Progress_Planner;
 
 use Progress_Planner\Update\Update_111;
 use Progress_Planner\Update\Update_130;
-
+use Progress_Planner\Update\Update_140;
 /**
  * Plugin Upgrade class.
  *
@@ -43,6 +43,7 @@ class Plugin_Migrations {
 	private const UPGRADE_CLASSES = [
 		'1.1.1' => Update_111::class,
 		'1.3.0' => Update_130::class,
+		'1.4.0' => Update_140::class,
 	];
 
 	/**
